Apply review changes
