+++ catalog/src/main/kotlin/com/adevinta/spark/catalog/examples/example/ExampleItem.kt
@@ -39,7 +39,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.adevinta.spark.catalog.examples.model.example.Example
+import com.adevinta.spark.catalog.examples.model.Example
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
